more doubles

+++ VTK/Graphics/vtkBlankStructuredGrid.cxx
@@ -46,7 +46,7 @@ vtkBlankStructuredGrid::~vtkBlankStructuredGrid()
 template <class T>
 void vtkBlankStructuredGridExecute(vtkBlankStructuredGrid *vtkNotUsed(self),
                                    T *dptr, int numPts, int numComp,
-                                   int comp, float min, float max,
+                                   int comp, double min, double max,
                                    vtkUnsignedCharArray *blanking)
 {
   T compValue;
